Support networking relink flow
