+++ financial-connections/src/main/java/com/stripe/android/financialconnections/repository/CoreAuthorizationPendingNetworkingRepairRepository.kt
@@ -3,15 +3,15 @@ package com.stripe.android.financialconnections.repository
 import android.os.Parcelable
 import androidx.lifecycle.SavedStateHandle
 import com.stripe.android.core.Logger
+import com.stripe.android.financialconnections.di.ActivityRetainedScope
 import com.stripe.android.financialconnections.repository.CoreAuthorizationPendingNetworkingRepairRepository.State
 import kotlinx.parcelize.Parcelize
 import javax.inject.Inject
-import javax.inject.Singleton
 
 /**
  * Repository for storing the core authorization pending repair.
  */
-@Singleton
+@ActivityRetainedScope
 internal class CoreAuthorizationPendingNetworkingRepairRepository @Inject constructor(
     savedStateHandle: SavedStateHandle,
     private val logger: Logger,
